refactor(vta-service): extract vta-keys (key management)
Third decomposition step (after vta-keyspaces+vta-vault #780, vta-config #781).
With config a crate, keys became a clean leaf — it reaches only into
vta-keyspaces, vta-config, vti-common, vti-secrets, vta-sdk, never into any
vta-service internal.

New vta-keys: master-seed storage, BIP-32 derivation, AES-GCM key wrapping,
imported-key handling, and seed-store backend selection, ~2.8k lines from
vta-service/src/keys/. Security-sensitive: seed material stays in
zeroize-guarded buffers. vta-service re-exports it as crate::keys (36 files
unchanged; vta_service::keys still resolves for vta-enclave). The eight
seed-store backend features chain to the matching vta-keys/* feature.

Pure move, verified hard: git renames (history preserved); only two pub(crate)
encoders (encode_private/public_multibase, used by operations::keys,
provision_integration, did_webvh) become pub at the boundary. 36 key tests in
the new crate; operations::keys (12) and the operations::backup
seed->derive->store->export->import->re-derive round-trip (39) pass unchanged;
vta-enclave builds with tee (compiles the KMS seed-store path). vta-service
drops ~2.8k lines (-> ~101.3k).

cargo test --workspace (default features, as CI runs): green. Clippy + fmt clean.
