+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
   "vta-cli-common",
   "vta-config",
   "vta-keyspaces",
+  "vta-keys",
   "vta-mcp",
   "vta-mobile-core",
   "vta-sdk",
@@ -32,6 +33,7 @@ default-members = [
   "vta-cli-common",
   "vta-config",
   "vta-keyspaces",
+  "vta-keys",
   "vta-mcp",
   "vta-sdk",
   "vta-service",
